+++ fire fighting robot/Fire Fighting Robot/DRIVERS/usart1.h
@@ -5,6 +5,6 @@
 #define USART1_DMA_RX_LEN 256
 #define USART1_DMA_TX_LEN 256
 void Usart1_Init(u32 baudrate);
-void Usart1_Send(unsigned char *DataToSend,u16 data_num);
+void Usart1_Send(unsigned char *DataToSend,u8 data_num);
 
 #endif
